@@ -1,4 +1,3 @@
-import React from "react";
 import { animated, useTrail } from "react-spring";
 import { useAppState } from "../AppStateProvider";
 import { useSelector } from "@xstate/react";
@@ -6,16 +5,14 @@ import { useSelector } from "@xstate/react";
 const DELAY_FACTOR = 500;
 
 interface AnimatedContentProps {
-  children: JSX.Element[] | JSX.Element;
+  children: React.ReactNode;
 }
 
 /**
  * Animates multiple elements onto the screen in a staggered pattern
- * @param children an array of JSX.Elements to animate
+ * @param children an array of elements to animate
  */
-export const AnimatedContent = ({
-  children,
-}: AnimatedContentProps): JSX.Element => {
+export const AnimatedContent = ({ children }: AnimatedContentProps) => {
   const content = Array.isArray(children) ? children : [children];
 
   const { appState, isReady } = useAppState();
@@ -24,8 +21,8 @@ export const AnimatedContent = ({
   const contentTrail = useTrail(content.length, {
     delay: DELAY_FACTOR,
     config: { mass: 25, tension: 2000, friction: 500 },
-    from: { y: "-15vh", x: "100vw", display: "none" },
-    to: introDone && { y: "0", x: "0", display: "block" },
+    from: { y: "-15vh", x: "100vw" },
+    to: introDone && { y: "0", x: "0" },
   });
 
   return (
