@@ -1,9 +1,8 @@
-import React from "react";
-import { Container } from "../../components/Container";
-import { AnimatedContent } from "../../components/AnimatedContent";
-import { Resume } from "../../components/Resume";
+import { Container } from "@/components/Container";
+import { AnimatedContent } from "@/components/AnimatedContent";
+import { Resume } from "@/components/Resume";
 
-export const ResumeScene = (): JSX.Element => {
+export const ResumeScene = () => {
   return (
     <Container>
       <AnimatedContent>
